requirements.txt
Package             Version    
------------------- -----------
backcall            0.1.0      
Brlapi              0.6.5      
Cartopy             0.17.0     
chardet             2.3.0      
cloudpickle         0.6.1      
cmocean             1.2        
cryptography        1.7.1      
cupshelpers         1.0        
cycler              0.10.0     
Cython              0.29.1     
dask                1.0.0      
decorator           4.3.0      
h5py                2.8.0      
h5pyd               0.3.3      
httplib2            0.9.2      
idna                2.2        
ipython             7.2.0      
ipython-genutils    0.2.0      
jedi                0.13.1     
keyring             10.1       
keyrings.alt        1.3        
kiwisolver          1.0.1      
louis               3.0.0      
Mako                1.0.6      
MarkupSafe          0.23       
matplotlib          3.0.2      
networkx            2.2        
numexpr             2.6.8      
numpy               1.15.4     
opencv-python       3.4.4.19   
pandas              0.23.4     
parso               0.3.1      
pexpect             4.2.1      
pickleshare         0.7.5      
Pillow              5.3.0      
pip                 18.1       
pkg-resources       0.0.0      
prompt-toolkit      2.0.7      
ptyprocess          0.5.1      
pvlib               0.6.0      
pyasn1              0.1.9      
pycrypto            2.6.1      
pycups              1.9.73     
pycurl              7.43.0     
Pygments            2.3.0      
pygobject           3.22.0     
pyparsing           2.3.0      
pyshp               2.0.1      
pysmbc              1.0.15.6   
pysolargeometry     0.1        
python-apt          1.4.0b3    
python-dateutil     2.7.5      
python-debian       0.1.30     
python-debianbts    2.6.1      
pytz                2018.7     
PyWavelets          1.0.1      
pyxdg               0.25       
reportbug           7.1.7      
reportlab           3.3.0      
requests            2.12.4     
scikit-image        0.14.1     
scikit-learn        0.20.1     
scipy               1.1.0      
SecretStorage       2.3.1      
setuptools          40.6.2     
Shapely             1.6.4.post2
six                 1.10.0     
sky-imager-api      0.2.0      
skysol              0.4        
speedtest-cli       2.0.2      
tables              3.4.4      
toolz               0.9.0      
traitlets           4.3.2      
unattended-upgrades 0.1        
urllib3             1.19.1     
utm                 0.4.2      
virtualenv          16.1.0     
wcwidth             0.1.7      
wheel               0.32.3     
youtube-dl          2017.5.18.1
